refactor: decompose useAIConversation into focused hooks (Task 2)
- Extract useAIStreaming for WebSocket event handlers
- Extract useAICodeApplication for code application logic
- Extract useAITimeout for request timeout management
- Refactor useAIConversation to orchestrate smaller hooks

Part of Issue #60
